requirements.txt
annotated-types==0.6.0
asgiref==3.7.2
Django==5.0.3
django-extensions==3.2.3
django-ninja==1.1.0
flake8==7.0.0
Jinja2==3.1.3
MarkupSafe==2.1.5
mccabe==0.7.0
mysqlclient==2.2.4
pipdeptree==2.16.1
ply==3.11
psycopg2-binary==2.9.9
pycodestyle==2.11.1
pydantic==2.6.4
pydantic_core==2.16.3
pyflakes==3.2.0
simple-ddl-parser==1.0.3
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.10.0
